pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "imx"
version = "0.1.0"
description = "CLI toolkit for image and video batch processing"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "opencv-python",
    "typer[all]",
    "natsort",
    "rich",
    "Pillow",
]

[project.scripts]
imx = "imx.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
